Image_Scripts/image_caption_generator/requirements.txt
tensorflow
tensorflow.keras
numpy
os
pickle
tqdm.notebook
nltk
PIL
matplotlib
